fix: logger removed from interpreter-funcrions throw errs
