+++ lib/app.js
@@ -3,7 +3,7 @@
 const program = require("commander");
 const init = require("./init.js").init;
 const utils = require("./utils.js");
-const logger = utils.logger;
+const logger = require("./logger.js");
 const encrypt = utils.encrypt;
 const path = require("path");
 const mongooseCloseConnection = utils.mongooseCloseConnection;
@@ -84,8 +84,8 @@ process.on("uncaughtException", (err) => {
   logger.log("error", err.stack);
 });
 
-process.on("unhandledRejection", (reason) => {
-  logger.log("error",reason);
+process.on("unhandledRejection", (reason, p) => {
+  logger.log("error",p,reason);
   process.exit();
 });
