+++ lib/classes/chain.js
@@ -3,7 +3,7 @@
 const utils = require("../utils.js");
 const replaceWithSmart = utils.replaceWithSmart;
 const dependsProcess = require("../depends-process.js");
-const logger = utils.logger;
+const logger = require("../logger.js");
 const getProcessByUId = utils.getProcessByUId;
 const chronometer = utils.chronometer;
 const crypto = require("crypto");
