+++ lib/classes/execution.js
@@ -3,7 +3,7 @@
 const utils = require("../utils.js");
 const ms = require("millisecond");
 const replaceWithSmart = utils.replaceWithSmart;
-const logger = utils.logger;
+const logger = require("../logger.js");
 const checkExecutorParams = utils.checkExecutorParams;
