+++ lib/classes/notification.js
@@ -3,7 +3,7 @@
 const utils = require("../utils.js");
 const replaceWithSmart = utils.replaceWithSmart;
 const checkNotifierParams = utils.checkNotifierParams;
-const logger = utils.logger;
+const logger = require("../logger.js");
 const qnm = require("../queue-notifications-memory.js");
 const qnr = require("../queue-notifications-redis.js");
 const crypto = require("crypto");
