+++ lib/classes/notificationEvent.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const utils = require("../utils.js");
-const logger = utils.logger;
+const logger = require("../logger.js");
 const loadConfigSection = utils.loadConfigSection;
 
 class notificationEvent {
