+++ lib/classes/process.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const utils = require("../utils.js");
-const logger = utils.logger;
+const logger = require("../logger.js");
 const loadConfigSection = utils.loadConfigSection;
 const replaceWithSmart = utils.replaceWithSmart;
 const getChainByUId = utils.getChainByUId;
