+++ lib/classes/queueSubscription.js
@@ -1,8 +1,7 @@
 "use strict";
 
-const utils = require("../utils.js");
+const logger = require("../logger.js");
 const crypto = require("crypto");
-const logger = utils.logger;
 let globalPlan = global.runtimePlan.plan;
 const configQueues = global.config.general.queues;
 const refreshInterval = configQueues.refreshInterval || 5000;
