+++ lib/classes/trigger.js
@@ -4,7 +4,7 @@ const utils = require("../utils.js");
 const queue = require("../queue-process-memory");
 const replaceWithSmart = utils.replaceWithSmart;
 const checkCalendar = utils.checkCalendar;
-const logger = utils.logger;
+const logger = require("../logger.js");
 
 class Trigger {
   constructor(chain, params) {
