+++ lib/init.js
@@ -1,8 +1,9 @@
 "use strict";
 
+
 const utils = require("./utils.js");
-const isUrl = require("./utils.js").isUrl;
-const logger = utils.logger;
+const isUrl = utils.isUrl;
+const logger = require("./logger.js");
 const path = require("path");
 const loadGeneralConfig = utils.loadGeneralConfig;
 const loadRemoteGeneralConfig = utils.loadRemoteGeneralConfig;
