+++ lib/interpreter-functions.js
@@ -4,7 +4,7 @@ const moment = require("moment");
 const crypto = require("crypto");
 const path   = require("path");
 const url    = require("url");
-const logger = require("./utils.js").logger;
+
 
 const functions = {
   gv: getvalue,
@@ -98,24 +98,27 @@ function getENVs(s) {
  * @returns {string}
  */
 function getvalue(idvalue, quote, values) {
-  let match = values[idvalue];
   let res = "";
+  if(values[idvalue]){
+    let match = values[idvalue];
 
-  if(match){
-    res = match;
+    if(match){
+      res = match;
+    }else{
+      if(idvalue.startsWith("ENV_")){
+        res = getENVs(idvalue);
+      }
+    }
+
+    if (quote){
+      if(quote === "\\'"){
+        quote = "'";
+      }
+      res = quote + res + quote;
+    }
   }else{
-    if(idvalue.startsWith("ENV_")){
-      res = getENVs(idvalue);
-    }
+    throw `value for value id "${idvalue}" not found.`;
   }
-
-  if (quote){
-    if(quote === "\\'"){
-      quote = "'";
-    }
-    res = quote + res + quote;
-  }
-
   return res;
 }
 
@@ -620,7 +623,7 @@ function pathParse(_path, property) {
   if(pathProperties[property]){
     res = pathProperties[property];
   }else{
-    logger.log("warn", `PathParse (${_path}) wrong property ${property}.`);
+    throw `PathParse (${_path}) wrong property ${property}.`;
   }
   return res;
 }
@@ -639,7 +642,7 @@ function urlParse(_url, property) {
   if(urlProperties[property]){
     res = urlProperties[property];
   }else{
-    logger.log("warn", `UrlParser (${_url}) wrong property ${property}.`);
+    throw `UrlParser (${_url}) wrong property ${property}.`;
   }
   return res;
 }
@@ -702,14 +705,12 @@ function quote(s, quote = "'") {
  */
 function stringify(obj) {
   if (!obj) {
-    logger.log("debug", `Interpreter: stringifying non value variable -> ${obj}`);
-    return "";
+    throw `stringifying non value variable -> ${obj}`;
   } else {
     try {
       return JSON.stringify(obj);
     } catch (ex) {
-      logger.log("error", `Interpreter: Stringify exception: ${ex}`);
-      return "";
+      throw `Stringify exception: ${ex}`;
     }
   }
 }
