+++ lib/interpreter.js
@@ -251,10 +251,9 @@ function evaluate(parseTree, values) {
 
       // GETVALUE / GETVALUEESCAPE / GETVALUEUNESCAPE:
       let _name = node.name.toLowerCase().substring(1);
-      if(["gv","getvalue","gvescape","getvalueescape","gvunescape","getvalueunescape"].indexOf(_name) !== -1){
+      if(["gvescape","getvalueescape","gvunescape","getvalueunescape"].indexOf(_name) !== -1){
         node.args[1] = values;
-      }
-      if(["gvq","getvaluequoted"].indexOf(_name) !== -1){
+      }else if(["gv","getvalue","gvq","getvaluequoted"].indexOf(_name) !== -1){
         node.args[2] = values;
       }
