+++ lib/queues.js
@@ -1,8 +1,7 @@
 "use strict";
 
 const queueSubscription = require("./classes/queueSubscription.js");
-const utils = require("./utils.js");
-const logger = utils.logger;
+const logger = require("./logger.js");
 
 let queues = ["zero"];
 global.processQueues = [];
