+++ lib/utils.js
@@ -1,6 +1,5 @@
 "use strict";
 
-const winston = require("winston");
 const fs = require("fs");
 const path = require("path");
 const configSchema = require("./schemas/config.json");
@@ -13,7 +12,6 @@ const redis = require("redis");
 const mongoose = require("mongoose");
 const lodash = require("lodash");
 const request = require("request");
-const interpret = require("./interpreter.js");
 const queue = require("./queue-process-memory.js");
 // Servers
 const express = require("express");
@@ -21,17 +19,11 @@ const bodyParser = require("body-parser");
 const app = express();
 const http = require("http");
 const https = require("https");
+const logger = require("./logger.js");
+const interpret = require("./interpreter.js");
+
 
 const algorithm = "aes-256-ctr";
-
-const debugMode = (process.env.RUNNERTY_DEBUG == "true");
-const testMode = (process.env.RUNNERTY_TEST == "true");
-
-const logger = new (winston.Logger)({
-  transports: [
-    new (winston.transports.Console)({colorize:(testMode)?false:"all", level: (debugMode)?"debug":"info"})
-  ]
-});
 
 module.exports.logger = logger;
