+++ package.json
@@ -40,8 +40,9 @@
     "lodash": "^4.17.10",
     "millisecond": "^0.1.2",
     "moment": "^2.22.1",
-    "mongoose": "^5.0.16",
+    "mongoose": "^5.1.3",
     "morgan": "^1.9.0",
+    "natives": "^1.1.4",
     "object-sizeof": "^1.2.0",
     "redis": "^2.8.0",
     "request": "^2.85.0",
@@ -81,9 +82,9 @@
     "commitizen": "^2.9.6",
     "cz-conventional-changelog": "^2.1.0",
     "eslint": "^4.19.1",
-    "eslint-plugin-jest": "^21.15.1",
-    "jest": "^22.4.3",
-    "jest-cli": "^22.4.3"
+    "eslint-plugin-jest": "^21.17.0",
+    "jest": "^23.0.1",
+    "jest-cli": "^23.0.1"
   },
   "jest": {
     "testEnvironment": "node",
